Small bug fix to chi sq calc in damocles code
Previously the rebinned model flux and flux error were scaled for the uneven frequency bin distribution before being
written out. But this scale was applied after the chi sq calculation was run. I've fixed it so the scaling is applied before
the chi sq calculation and before the arrays are written out in integrated_line_profile_binned.out